texk/dvipdfm-x: Minor bug fix and cleanup
git-svn-id: svn://tug.org/texlive/trunk/Build/source/texk/dvipdfm-x@35246 c570f23f-e606-0410-a88d-b1316a301751

Conflicts:
	ChangeLog
	bmpimage.c
	bmpimage.h
	jp2image.c
	jp2image.h
	xbb.c

+++ bmpimage.c
@@ -88,7 +88,7 @@ get_density (double *xdensity, double *ydensity, struct hdr_info *hdr)
 }
 
 int
-texpdf_bmp_get_bbox (FILE *fp, long *width, long *height,
+texpdf_bmp_get_bbox (FILE *fp, int *width, int *height,
               double *xdensity, double *ydensity)
 {
   int r;
